create user model
Here we have created a folder called api and in there is an another folder called models which contains user_model. the user_model contains information on how we can create userSchema adding the necessary type,unique,and required info and we created timestamps-true to create two extra creation of time(time of creation and time of edit) . and finally the mongoose.model function is used to create a model named "User" using the "userSchema" . This model represents a schema for a user document in a MongoDB collection.